feat: Add password reset
You can now use the "FORGOT PASSWORD" button on the login page to send
a password reset email to the email in the email field. In that email,
you are given a link, through which you get to a site that will prompt
you for a new password, and then request the server to change the
password hash in the database. After that, you can login using the new
password.